requirements.txt
libigl
polyscope
numpy
scipy
matplotlib
scikit-learn